Implement Compression Round functionality for buttons
Buttons added to allow for input of new message for compression
and to control compression cycle allowing only 1 round to be completed
at a time.  SHA_256 method was separated into individual methods to allow
for one round completion with each button click.  This will allow
compression round values to be viewed on screen when labels are clicked.