Tutor Logic/Application Navigation: Add a tutor recommendation Button.
The logic looks right.  TutoringSessionView should display a Request Problem button next to the Go to Dashboard button.  Should be able to check if assessment type has been completed with hasCompleted, and what the student is weakest in with getWeakestProblemType found in StudentModel.  Logic for button Request Problem is in ShaTuTutor is using functions hasCompleted and getWeakestProblemType.  References added to KnowledgeComponent and ServiceFactory as needed.